+++ knip.config.ts
@@ -1,0 +1,4 @@
+export default {
+  entry: ["packages/**/index.{js,ts}", "packages/**/src/index.{js,ts}"],
+  project: ["**/*.{js,ts}"],
+};
