+++ package.json
@@ -6,13 +6,18 @@
     "task": "bun run examples/example-basic.ts",
     "twitter": "bun run examples/example-twitter.ts",
     "test": "bun run packages/core",
-    "build": "pnpm --dir packages/core run build"
+    "build": "pnpm --dir packages/core run build",
+    "knip": "knip"
   },
   "dependencies": {
-    "@types/node": "^22.10.5",
     "ajv": "^8.17.1",
     "chalk": "^5.4.1",
     "readline": "^1.3.0",
     "zod": "^3.24.1"
+  },
+  "devDependencies": {
+    "@types/node": "^22.10.5",
+    "knip": "^5.43.1",
+    "typescript": "~5.6.2"
   }
 }
