+++ packages/core/src/core/index.ts
@@ -1,0 +1,35 @@
+import { Core } from "./core";
+import { RoomManager } from "./room-manager";
+import { Room } from "./room";
+import { ChromaVectorDB } from "./vector-db";
+import { Processor } from "./processor";
+import { GoalManager } from "./goal-manager";
+import { ChainOfThought } from "./chain-of-thought";
+import { TaskScheduler } from "./task-scheduler";
+import { Logger } from "./logger";
+import { LogLevel } from "../types";
+import { Consciousness } from "./consciousness";
+import { LLMClient } from "./llm-client";
+import { StepManager } from "./step-manager";
+import { defaultCharacter } from "./character";
+import * as Utils from "./utils";
+import * as Providers from "./providers";
+
+export {
+  Core,
+  Consciousness,
+  LLMClient,
+  StepManager,
+  TaskScheduler,
+  Logger,
+  LogLevel,
+  RoomManager,
+  Room,
+  ChromaVectorDB,
+  Processor,
+  GoalManager,
+  ChainOfThought,
+  Utils,
+  defaultCharacter,
+  Providers,
+};
