+++ packages/core/src/core/vector-db.ts
@@ -3,137 +3,23 @@ import { ChromaClient, IncludeEnum, OpenAIEmbeddingFunction } from "chromadb";
 import { env } from "./env";
 import { Logger } from "./logger";
 import { Room } from "./room";
-import { LogLevel, type SearchResult } from "../types";
-
-// ======================= INTERFACES & TYPES =======================
-
-export interface VectorDB {
-  findSimilar(
-    content: string,
-    limit?: number,
-    metadata?: Record<string, any>
-  ): Promise<SearchResult[]>;
-
-  store(content: string, metadata?: Record<string, any>): Promise<void>;
-
-  delete(id: string): Promise<void>;
-
-  storeInRoom?(
-    content: string,
-    roomId: string,
-    metadata?: Record<string, any>
-  ): Promise<void>;
-
-  findSimilarInRoom?(
-    content: string,
-    roomId: string,
-    limit?: number,
-    metadata?: Record<string, any>
-  ): Promise<SearchResult[]>;
-
-  storeSystemMetadata(key: string, value: Record<string, any>): Promise<void>;
-  getSystemMetadata(key: string): Promise<Record<string, any> | null>;
-
-  storeEpisode(memory: Omit<EpisodicMemory, "id">): Promise<string>;
-  findSimilarEpisodes(
-    action: string,
-    limit?: number
-  ): Promise<EpisodicMemory[]>;
-  getRecentEpisodes(limit?: number): Promise<EpisodicMemory[]>;
-
-  storeDocument(doc: Omit<Documentation, "id">): Promise<string>;
-  findSimilarDocuments(query: string, limit?: number): Promise<Documentation[]>;
-  searchDocumentsByTag(
-    tags: string[],
-    limit?: number
-  ): Promise<Documentation[]>;
-  updateDocument(id: string, updates: Partial<Documentation>): Promise<void>;
-
-  purge(): Promise<void>;
-}
-
-export interface EpisodicMemory {
-  id: string;
-  timestamp: Date;
-  action: string;
-  outcome: string;
-  context?: Record<string, any>;
-  emotions?: string[];
-  importance?: number;
-}
-
-export interface Documentation {
-  id: string;
-  title: string;
-  content: string;
-  category: string;
-  tags: string[];
-  lastUpdated: Date;
-  source?: string;
-  relatedIds?: string[];
-}
-
-interface Cluster {
-  id: string;
-  name: string;
-  description: string;
-  centroid?: number[];
-  topics: string[];
-  documentCount: number;
-  lastUpdated: Date;
-}
-
-interface ClusterMetadata {
-  clusterId: string;
-  confidence: number;
-  topics: string[];
-}
-
-interface ClusterStats {
-  variance: number;
-  memberCount: number;
-  averageDistance: number;
-}
-
-interface ClusterUpdate {
-  newCentroid?: number[];
-  documentCount: number;
-  topics: string[];
-  variance?: number;
-}
-
-interface DocumentClusterMetadata extends ClusterMetadata {
-  category: string;
-  commonTags: string[];
-}
-
-interface EpisodeClusterMetadata extends ClusterMetadata {
-  commonEmotions: string[];
-  averageImportance: number;
-}
-
-interface HierarchicalCluster extends Cluster {
-  parentId?: string;
-  childIds: string[];
-  level: number;
-  domain: string;
-  subDomain?: string;
-}
-
-interface DomainMetadata {
-  domain: string;
-  subDomain?: string;
-  confidence: number;
-}
+import {
+  LogLevel,
+  type ClusterMetadata,
+  type ClusterStats,
+  type ClusterUpdate,
+  type Documentation,
+  type DocumentClusterMetadata,
+  type DomainMetadata,
+  type EpisodeClusterMetadata,
+  type EpisodicMemory,
+  type HierarchicalCluster,
+  type SearchResult,
+  type VectorDB,
+} from "../types";
+import { isValidDateValue } from "./utils";
 
 // Helper function to check if a value can be used for a Date constructor
-function isValidDateValue(value: unknown): value is string | number | Date {
-  return (
-    typeof value === "string" ||
-    typeof value === "number" ||
-    value instanceof Date
-  );
-}
 
 // ======================= MAIN CLASS =======================
